Create code_this
using the HTML basic lesson on Sololearn, I created this. the other users in the comments section are a great help.
i clicked on every "try it yourself" in the lesson, which directs users to the "code playground" & copied the code, pasted to notepad, & edited with Code Writer app. there were many redundant tags which i removed.